+++ src/Finis.Application/Dto/CompraAtivo/CompraAtivoView.cs
@@ -11,7 +11,7 @@ public class CompraAtivoView
     public decimal ValorCompra { get; set; }
     public decimal EstimativaVenda { get; set; }
     public decimal ValorCota { get; set; }
-    public int QtdCotas { get; set; }
+    public decimal QtdCotas { get; set; }
     public bool FlVendido { get; set; }
     public required string FlBolsa { get; set; }
     public required string Corretora { get; set; }
