+++ src/Finis.Domain/Models/CompraAtivo.cs
@@ -12,7 +12,7 @@ public class CompraAtivo
     public decimal ValorCompra { get; set; }
     public decimal EstimativaVenda { get; set; }
     public decimal ValorCota { get; set; }
-    public int QtdCotas { get; set; }
+    public Decimal QtdCotas { get; set; }
     public bool FlVendido { get; set; }
     public required string FlBolsa { get; set; }
     public required string Corretora { get; set; }
